admin后台中announcement模块的开发 1. 实现了新建announcement的基本功能 2. 添加了StringUtils功能 3. 实现了输入校验等

+++ src/main/java/com/lkk/lostfound/dao/DaoBase.java
@@ -6,7 +6,7 @@ public interface DaoBase<T> {
 
 	T get(long entityId);
 
-	Integer save(T entity);
+	Long save(T entity);
 
 	void update(T entity);
